Progress on analog input monitoring.
Currently using NI DAQmx callbacks to read data when it is available,
but this is finnicky - only accepts reading certain numbers of points
at a time in a way that is difficult to predict. Going to switch to
acquiring in a thread with a short timeout and see if that is better.