Scanning going pretty well, but a hitch:
It's pretty impossible to not have all the tasks very coupled together.
The counter task needs a clock, which is either from the AI sample
clock or the AO sample clock depending on whether we're monitoring or
scanning. And you can't have multiple AI tasks using different AI
terminals.

So we'll have to merge our methods into a single start_tasks method and
a single read_tasks method that are sufficiently flexible to  run
regardless of whether we are in monitoring or scanning mode.