Catch DAQmx errors starting tasks
This is how we're catching invalid settings for now - relying on DAQmx
throwing an error instead of validating it ourselves.

Abort transitioning to RUNNING or SCANNING if a DAQError occurs,
clean up, and show an error dialog box.

Also removed fake data injection. We're ready for testing on the
apparatus.